refactor: 권한 확인 및 이름 중복 검사 추가
